requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anyio==4.5.2
async-timeout==5.0.1
attrs==25.1.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
dateparser==1.2.1
exceptiongroup==1.2.2
fonttools==4.56.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.5.3
multidict==6.1.0
numpy==1.21.6
packaging==24.2
pandas==1.3.5
pillow==10.4.0
propcache==0.2.0
pycryptodome==3.21.0
pyparsing==3.1.4
python-binance==1.0.16
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-telegram-bot==13.15
pytz==2025.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.0.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.2.3
websockets==13.1
yarl==1.15.2
zipp==3.20.2
tensorflow==2.10.1
keras==2.10.0
ta==0.10.2
tqdm==4.64.1